fix: Ensure content updates don't cause duplication
Content is initially loaded from DB but updated if we hear back from the
network. It is important to check for duplicates.

In the case of neighborhood and cuisine dropdowns this can be done
directly since these are easy to compare (just strings). For other
objects, e.g. restaurants and reviews, it is not so easy to compare.
Thus for now the following objects are purged before populating from
JSON:
* restaurant list
* details page:
  * restaurant image
  * restaurant hours
  * restaurant reviews