fix: Ensure selects are populated before restaurant list
If selects where not at their default values before navigating away then
navigating back with the browser back button raised an error.

This was because the select values were accessed before the selects were
populated (and possibly Chrome attempted to set the previous values).
The solution was changing the order of calls onload and using await to
enforce syncronicity.