refactor: Create page controller classes
* `index.html` controlled by `indexController`
* `restaurant_info.html` controlled by `restaurantInfoController`

Also created helper classes for restaurant and review data to handle
HTML generation and auxiliary tasks:
* `class Restaurant`
* `class Review`
This classes are held in their own files which are loaded as modules by
the browser (ooh shiny!).

Finally, `/js/sw.js` has been removed since the filename is confusing −
it implies it is the service worker. In fact it was the service worker
loading code. This code is now included inline in each HTML file in a
`<script>` tag.