fix: Stop dropping Restaurant and Review objects
Only restaurants/reviews not already present in the DOM are inserted.
Before, however, the old Restaurant/Review objects were thrown away and
new ones created. In particular this caused a bug where all restaurant
map pins appeared on the map at all times. The pointers to the pins were
thrown away and new pins drawn on top. It was thus impossible to clear
the old pins.

Now the old Restaurant/Review objects are kept and objects for any new
items appended. This fixes the issue.

Note, that it remains the case that update to restaurants/reviews
already shown on th page will not be applied.