feat: Notify of pending transactions
Add `_isPending()` function to DBHelper which returns (promise resolving
to) a boolean. This value determines if there are pending transactions
waiting to be sent over the network.

DBHelper now accepts an optional callback `pendingCallback`, which is
called with `false` when transactions are cleared successfully, and
`true` when the first pending transaction is added.